move whoami.ts into user/
